fix(obstacle_segmentation): use modified rosbag on obstacle segmentation
